fix: fix the status code and upload logo in update state system

+++ state/controllers/stateRegistration.controller.js
@@ -17,9 +17,9 @@ exports.getRegistration = async (req, res)=>{
                 result = await stateRegistration.query().where({stateID});
             }
             else{
-                result = {
-                    message: "Maaf selain panitia tidak diperkenankan untuk mengaksesnya"
-                }
+                return res.status(403).send({
+                    message: "Maaf selain panitia, tidak diperkenankan untuk mengaksesnya."
+                })
             }
 
         }
@@ -32,9 +32,9 @@ exports.getRegistration = async (req, res)=>{
                 result = await stateRegistration.query().where({stateID, nim});
             }
             else{
-                result = {
-                    message: "Maaf selain panitia tidak diperkenankan untuk mengaksesnya"
-                }
+                return res.status(403).send({
+                    message: "Maaf selain panitia, tidak diperkenankan untuk mengaksesnya."
+                })
             }
 
         }
@@ -42,7 +42,7 @@ exports.getRegistration = async (req, res)=>{
         return res.status(200).send(result);
     }
     catch(err){
-        return res.status(400).send({
+        return res.status(500).send({
             message: err.message
         });
     }
@@ -97,7 +97,7 @@ exports.attendanceState = async (req, res)=>{
         });
 
         if(checkRegistration.length === 0){
-            return res.status(400).send({message: "Anda belum mendaftar"});
+            return res.status(404).send({message: "Anda belum mendaftar"});
         }
 
         const updateResult = await stateRegistration.query()
@@ -127,7 +127,7 @@ exports.updateAttendance = async (req, res)=>{
         });
 
         if(checkRegistration.length === 0){
-            return res.status(400).send({message: "Peserta belum mendaftar"});
+            return res.status(404).send({message: "Peserta belum mendaftar"});
         }
 
         const updateAttendance = await stateRegistration.query().where({
@@ -152,23 +152,36 @@ exports.verifyAttendanceCode = async (req, res) => {
     const {nim} = req.query;
     const {attendanceCode} = req.body;
     const exitAttendance = 1;
-
+    
     try{
-        const stateAttendanceCode = await stateActivities.query()
-        .select('attendanceCode')
-        .where('stateID', stateID);
+        const stateAttendanceCode = await stateRegistration.query()
+        .select('state_activities.attendanceCode')
+        .where({
+            'state_registration.stateID': stateID,
+            'state_registration.nim': nim
+        })
+        .join(
+            'state_activities', 
+            'state_activities.stateID', 
+            'state_registration.stateID'
+        );
+
+        if(stateAttendanceCode.length === 0)
+            return res.status(404).send({
+                message: "Anda belum mendaftar"
+            });
+        
         
         if(attendanceCode === stateAttendanceCode[0].attendanceCode){
             const updateResult = await stateRegistration.query()
             .patch({exitAttendance})
             .where({stateID,nim});
 
-            return res.status(200).send({message: "Anda Sudah Absen"});
+            return res.status(200).send({message: "Proses Absensi Selesai"}); 
         }
         else{
-            return res.status(500).send({message: "Kode presensi salah"});
+            return res.status(406).send({message: "Kode presensi salah"});
         }
-
     }
     catch(err){
         return res.status(500).send({message: err.message});
@@ -180,6 +193,14 @@ exports.deleteRegistration = async (req, res)=>{
     const {nim} = req.query;
     
     try{
+        const checkRegistration = await stateRegistration.query()
+        .where({nim, stateID});
+
+        if(checkRegistration.length === 0)
+            return res.status(404).send({
+                message: "Anda belum mendaftar"
+            });
+
         const deleteResult = await stateRegistration.query()
         .delete()
         .where({nim, stateID});
@@ -187,7 +208,7 @@ exports.deleteRegistration = async (req, res)=>{
         const registeredState = await stateActivities.query().select('registered');
 
         const updateRegisteredState = await stateActivities.query()
-        .where('stateID', stateID)
+        .where({stateID})
         .patch({
             registered: registeredState[0].registered - 1
         });
@@ -195,7 +216,7 @@ exports.deleteRegistration = async (req, res)=>{
         return res.status(200).send({message: "Data Registrasi Berhasil Dihapus"});
     }
     catch(err){
-        return res.status(400).send({message: err.message});
+        return res.status(500).send({message: err.message});
     }
 }
